+++ app/shared/io/helpers/implementations/HashMapWithArrayBufferAdderConcreteImplementation.scala
@@ -1,4 +1,4 @@
-package shared.io.helpers.implementation
+package shared.io.helpers.implementations
 
 import java.util.concurrent.ConcurrentMap
