+++ app/shared/io/helpers/implementations/NiceObject.scala
@@ -1,4 +1,4 @@
-package shared.io.helpers.implementation
+package shared.io.helpers.implementations
 
 class NiceObject[T <: AnyRef](x : T) {
   def niceClass : Class[_ <: T] = x.getClass.asInstanceOf[Class[T]]
