+++ app/shared/io/helpers/traits/MapHelperBase.scala
@@ -1,6 +1,6 @@
 package shared.io.helpers.traits
 
-import shared.io.helpers.implementation.HashMapWithArrayBufferAdderConcreteImplementation
+import shared.io.helpers.implementations.HashMapWithArrayBufferAdderConcreteImplementation
 
 trait MapHelperBase {
   lazy val hashMapWithArrayBufferAdder : HashMapWithArrayBufferAdder = new HashMapWithArrayBufferAdderConcreteImplementation
