+++ app/shared/io/helpers/traits/reflection/ClassTagHelperGetMethodWrappersAsMap.scala
@@ -3,7 +3,7 @@ package shared.io.helpers.traits.reflection
 import shared.com.ortb.model.reflection
 import shared.com.ortb.model.reflection.MethodWrapperModel
 import shared.com.ortb.model.results.ResultWithCountSuccessModel
-import shared.io.helpers.implementation.reflection.ClassTagHelperConcreteImplementation
+import shared.io.helpers.implementations.reflection.ClassTagHelperConcreteImplementation
 import shared.io.helpers.{ EmptyValidateHelper, MapHelper }
 
 import scala.collection.mutable.ArrayBuffer
