+++ app/shared/io/helpers/traits/reflection/ClassTagHelperMembers.scala
@@ -1,6 +1,6 @@
 package shared.io.helpers.traits.reflection
 
-import shared.io.helpers.implementation.reflection.ClassTagHelperConcreteImplementation
+import shared.io.helpers.implementations.reflection.ClassTagHelperConcreteImplementation
 
 trait ClassTagHelperMembers
   extends ClassTagHelperMemberInfoGetter
